Add support to define a prefix for services of each connection
By watching over skupper-prefixes configmap a map from "origin" id of
the connection to the desired prefix string, the prefix will be added
to the name of the created service (and the proxy deployment).

For now, we suppose that the prefix will be set before annotating the
services and will not be updated later.